% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscapeTest.R
\name{scene}
\alias{scene}
\title{<Add Title>}
\usage{
scene(data, width = NULL, height = NULL, elementId = NULL)
}
\description{
<Add Description>
}
\examples{
scene(mtcars) |>
  add_plot_wrapper("scatter", list(x = "wt", y = "mpg")) |>
  add_plot_wrapper("histo", list(x = "disp", y = "_indicator"))
}
